Tentative event edit added

+++ organizer/views.py
@@ -1,3 +1,4 @@
+from django.forms import ValidationError
 from django.shortcuts import redirect, render
 from organizer.models import Account, Category, Event
 from django.contrib.auth.decorators import login_required
@@ -7,6 +8,8 @@ from django.views.generic import TemplateView
 from django.utils import timezone
 from calendar import monthrange
 from datetime import timedelta
+from django.db import IntegrityError
+
 
 # id of the current logged in user (to make passing data btwn views easier)
 logged_in_user:int = 0
@@ -53,17 +56,22 @@ def settings(request):
         email = request.POST.get('email_text_box')
         category_name = request.POST.get('myCategories')
         password = request.POST.get('password_text_box')
+        delete_cat = request.POST.get('delete_category') # is None when unchecked, Delete when checked
+        
         color = request.POST.get('colors')
-        if category_name and color:
+        if delete_cat == 'Delete':
+            Category.objects.filter(name = category_name).delete()
+            messages.info(request, "category deleted!")
+        elif category_name and color:
             Category.objects.update_or_create(name = category_name, account_id_id = user.account_id, defaults ={'color':color})
             messages.info(request, "category info saved!")
-        if username and username != user.username:
+        elif username and username != user.username:
             Account.objects.update_or_create(account_id = user.account_id, defaults ={'username':username})
             messages.info(request, "updated username")
-        if password and password != user.password:
+        elif password and password != user.password:
             Account.objects.update_or_create(account_id = user.account_id, defaults ={'password':password})
             messages.info(request, "updated password")
-        if email and email != user.email:
+        elif email and email != user.email:
             Account.objects.update_or_create(account_id = user.account_id, defaults ={'email':email})
             messages.info(request, "updated email")
     data = Account.objects.values('email','password', 'created_on', 'category__name', 'category__color').filter(account_id = logged_in_user).distinct()
@@ -88,20 +96,38 @@ def events(request):
     categories = Category.objects.filter(account_id = logged_in_user).order_by("name")
     return render(request, "organizer/events.html", context={'events': events, 'acct' : logged_in_user, 'categories': categories})
 
-def command(request, id, cmd):
+def command(request, id:int, cmd:str):
+    global logged_in_user
+    if logged_in_user == 0:
+        return render(request, "organizer/login.html", context=None)
     event = Event.objects.get(pk = id)
-    acct = logged_in_user
     if cmd == "delete":
         event.delete()
     if cmd == "details":
-        return render(request, "organizer/event_detail.html", context={'event' : event})
-    #return render(request, "organizer/events.html", context={'events' : Event.objects.filter(account_id = acct).order_by("event_start_date")})
+        acct_id = logged_in_user
+        cats = Category.objects.filter(account_id = acct_id).order_by("name")
+        if request.method == 'POST':
+            acct = Account.objects.get(pk=acct_id)
+            category_id = request.POST.get('category_box')
+            category = Category.objects.get(pk = category_id)
+            location = request.POST.get('location_text_box')
+            event_start_date = request.POST.get('sdate')
+            event_start_time = request.POST.get('stime')
+            event_end_date = request.POST.get('edate')
+            event_end_time = request.POST.get('etime')
+            name = request.POST.get('n')
+            description = request.POST.get('description')
+            try:
+                Event.objects.filter(event_id = event.event_id).update(account_id = acct, category_id = category, location = location, event_start_date = event_start_date, event_start_time = event_start_time if event_start_time != '' else None, event_end_date = event_end_date if event_end_date != '' else None, event_end_time = event_end_time if event_end_time != '' else None, name = name, description = description)
+            except IntegrityError as e:
+                return render(request, "organizer/edit_event.html", context={"acct": acct_id, "success_message": "Error: " + str(e), "cats": cats, "event":event})
+            else:
+                return render(request, "organizer/edit_event.html", context={"acct": acct_id, "success_message": "updated!", "cats": cats, "event":event})#, "initial_event":event})
+        return render(request, 'organizer/edit_event.html', context = {'acct' : acct_id, 'cats': cats, "event":event})
     return redirect('events')
-
 
 class WeeklyCalendarView(TemplateView):
     template_name = 'organizer/calendar.html'
-
 
     def get_context_data(self, month = None, year = None, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -130,3 +156,32 @@ class WeeklyCalendarView(TemplateView):
         context['month_num'] = month
         context['dates'] = dates
         return context
+
+def addEvent(request):
+    global logged_in_user
+    if logged_in_user == 0:
+        return render(request, "organizer/login.html", context=None)
+    acct_id = logged_in_user
+    cats = Category.objects.filter(account_id = acct_id).order_by("name")
+    if request.method == 'POST':
+        acct = Account.objects.get(pk=acct_id)
+        category_id = request.POST.get('category_box')
+        category = Category.objects.get(pk = category_id)
+        location = request.POST.get('location_text_box')
+        event_start_date = request.POST.get('sdate')
+        event_start_time = request.POST.get('stime')
+        event_end_date = request.POST.get('edate')
+        event_end_time = request.POST.get('etime')
+        name = request.POST.get('n')
+        description = request.POST.get('description')
+        try:
+            print("test")
+            Event.objects.create(account_id = acct, category_id = category, location = location, event_start_date = event_start_date, event_start_time = event_start_time if event_start_time != '' else None, event_end_date = event_end_date if event_end_date != '' else None, event_end_time = event_end_time if event_end_time != '' else None, name = name, description = description)
+        except IntegrityError as e:
+            return render(request, "organizer/add_event.html", context={"acct": acct_id, "success_message": "Error: " + str(e), "cats": cats})
+        except ValidationError as e:
+            return render(request, "organizer/add_event.html", context={"acct": acct_id, "success_message": "check inputs", "cats": cats})
+
+        else:
+            return render(request, "organizer/add_event.html", context={"acct": acct_id, "success_message": "event created successfully!", "cats": cats})#, "initial_event":event})
+    return render(request, 'organizer/add_event.html', context = {'acct' : acct_id, 'cats': cats})
